style: fix Profile navigation header
