+++ src/components/Profile/Profile.jsx
@@ -1,15 +1,43 @@
 import { useState } from "react";
 import UserInfo from "./UserInfo";
-import { Outlet } from "react-router-dom";
+import { Outlet, useNavigate, NavLink } from "react-router-dom";
 import NoUser from "./NoUser";
+import { useSelector } from "react-redux";
 
 function Profile({ username }) {
     const [status, setStatus] = useState("user");
+    const otherProfileData = useSelector((state) => state.otherProfile);
+    const navigate = useNavigate();
 
     if (status === "nouser") return <NoUser username={username} />;
 
     return (
         <div>
+            <div className="top flex px-2 sticky z-50 top-0 backdrop-blur-[400px] opacity-[100%]">
+                <NavLink
+                    className="left my-auto p-3 hover:bg-gray-200 rounded-full"
+                    onClick={() => navigate(-1)}
+                >
+                    <svg
+                        viewBox="0 0 24 24"
+                        aria-hidden="true"
+                        className="w-5 m-auto r-4qtqp9 r-yyyyoo r-dnmrzs r-bnwqim r-lrvibr r-m6rgpd r-z80fyv r-19wmn03"
+                    >
+                        <g>
+                            <path d="M7.414 13l5.043 5.04-1.414 1.42L3.586 12l7.457-7.46 1.414 1.42L7.414 11H21v2H7.414z"></path>
+                        </g>
+                    </svg>
+                </NavLink>
+
+                <div className="right ml-4">
+                    <p className="font-bold text-xl">
+                        {otherProfileData?.name || ""}
+                    </p>
+                    <p className="text-[13px] my-0.5 font-light">
+                        {otherProfileData?.tweets || "0"} posts
+                    </p>
+                </div>
+            </div>
             <UserInfo username={username} setStatus={setStatus} />
 
             <Outlet />
